+++ libraries/base/changelog.md
@@ -130,6 +130,10 @@
 
   * Make `read . show = id` for `Data.Fixed` (#9240)
 
+  * Add `calloc` and `callocBytes` to `Foreign.Marshal.Alloc`.
+
+  * Add `callocArray` and `callocArray0` to `Foreign.Marshal.Array`.
+
 ## 4.7.0.2  *Dec 2014*
 
   * Bundled with GHC 7.8.4